ZezeJava: BeanFactory增加Data类的支持; 为了方便通用处理, Serializable接口增加typeId方法

+++ ZezeJava/ZezeJava/src/main/java/Zeze/Dbh2/LogPrepareBatch.java
@@ -23,7 +23,7 @@ public class LogPrepareBatch extends Log {
 	}
 
 	@Override
-	public int typeId() {
+	public long typeId() {
 		return TypeId_;
 	}
