+++ ZezeJava/ZezeJava/src/main/java/Zeze/Dbh2/LogSetSplittingMeta.java
@@ -21,7 +21,7 @@ public class LogSetSplittingMeta extends Log {
 	}
 
 	@Override
-	public int typeId() {
+	public long typeId() {
 		return TypeId_;
 	}
