+++ ZezeJava/ZezeJava/src/main/java/Zeze/Raft/RaftLog.java
@@ -62,7 +62,7 @@ public final class RaftLog implements Serializable {
 	public void encode(ByteBuffer bb) {
 		bb.WriteLong(term);
 		bb.WriteLong(index);
-		bb.WriteInt4(log.typeId());
+		bb.WriteInt4((int)log.typeId());
 		log.encode(bb);
 	}
